+++ src/frontend/apps/hub/src/features/auth/components/LogoutButton.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@gouvfr-lasuite/cunningham-react";
+import { Button } from "@gouvfr-lasuite/ui-components";
 import { logout } from "../Auth";
 import { useTranslation } from "react-i18next";