+++ src/frontend/apps/hub/src/features/config/ConfigProvider.tsx
@@ -1,4 +1,4 @@
-import { Spinner } from "@gouvfr-lasuite/ui-kit";
+import { Spinner } from "@gouvfr-lasuite/ui-components";
 import Head from "next/head";
 import Script from "next/script";
 import { createContext, useContext } from "react";