+++ src/frontend/apps/hub/src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import { CunninghamProvider } from "@gouvfr-lasuite/ui-kit";
+import { CunninghamProvider } from "@gouvfr-lasuite/ui-components";
 import {
   MutationCache,
   Query,